# repo/api/openapi/v1.yaml
openapi: 3.1.0
info:
  title: KuberCloud ANI Core API
  version: 1.0.0
  description: |
    KuberCloud ANI Core — 基础设施平台 REST API 契约（唯一真实来源）。

    ## 文件范围与边界
    本文件（api/openapi/v1.yaml）仅包含 ANI Core 基础设施资源：
    instances（VM/Container/GPU/Sandbox/BM/K8s集群）、volumes/filesystems/objects/
    vector-stores、networks、auth/secrets/encryption、gpu-inventory/dpu-inventory、
    registry/metering/observability/notifications/audit 等。
    ANI Services 资源（模型仓库/推理/知识库/PaaS）在 api/openapi/services/v1.yaml。

    ## URL 前缀规范
    servers[0].url = "https://{host}/api/v1"（前缀在 spec 中定义，Gateway 不做路径变换）。
    版本升级时新建 api/openapi/v2.yaml，本文件路径保持稳定。

    ## SDK 生成（make gen-core-sdk）
    Go: github.com/kubercloud/ani-go | Python: pip install kubercloud-ani
    TypeScript: @kubercloud/ani-client | Java: com.kubercloud:ani-java

    ## 认证
    Bearer JWT（1小时有效）或 X-API-Key Header。
    tenant_id 从 JWT claims 提取，请求体中的 tenant_id 字段将被忽略。

    ## 错误格式
    `{ code: string, message: string, request_id: string, details?: object }`
    标准错误码：UNAUTHORIZED / FORBIDDEN / NOT_FOUND / CONFLICT / BAD_REQUEST /
    RATE_LIMIT_EXCEEDED / NOT_IMPLEMENTED / INTERNAL_ERROR

    ## 分页
    cursor 分页：`?limit=20&cursor=<token>` → `{ items: [], next_cursor: string | null }`

    ## 异步操作
    202 响应体返回 AsyncTask，Location Header 附任务 URL。
    轮询 GET /api/v1/tasks/{task_id} 或配置 Webhook。
servers:
  - url: https://{host}/api/v1
    variables:
      host:
        default: ani.internal
        description: ANI Gateway 地址（支持IP地址）

...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT AccessToken（有效期1小时，通过 /auth/token 获取）
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 长期 API Key（在 Console → 设置 → API Key 中创建）

# ── 公共 Schema ───────────────────────────────────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      properties:
        code:        { type: string, example: "MODEL_NOT_FOUND" }
        message:     { type: string, example: "模型 qwen2.5-72b 不存在" }
        request_id:  { type: string, example: "req_01j8xabc" }
        details:     { type: object, additionalProperties: true }

    # Cursor 分页（所有列表接口统一使用此格式）
...
    CoreDevProfileInfo:
      type: object
      description: "Core dev/local profile 标记；用于区分本地联调成功与真实 provider 执行成功。"
      required: [mode, provider, real_provider]
      properties:
        mode:          { type: string, enum: [local, real] }
        provider:      { type: string }
        real_provider: { type: boolean }
        reason:        { type: string, nullable: true }
...
    NetworkResourceState:
      type: string
      enum: [pending, available, failed, deleting, deleted]
    NetworkVPC:
      type: object
      required: [id, tenant_id, name, cidr, state, created_at, updated_at]
      properties:
        id:         { type: string }
        tenant_id:  { type: string }
        name:       { type: string }
        cidr:       { type: string, example: "10.20.0.0/16" }
        state:      { $ref: '#/components/schemas/NetworkResourceState' }
        reason:     { type: string, nullable: true }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    NetworkSubnet:
      type: object
      required: [id, tenant_id, vpc_id, name, cidr, state, created_at, updated_at]
      properties:
        id:         { type: string }
        tenant_id:  { type: string }
        vpc_id:     { type: string }
        name:       { type: string }
        cidr:       { type: string, example: "10.20.1.0/24" }
        gateway:    { type: string, nullable: true }
        state:      { $ref: '#/components/schemas/NetworkResourceState' }
        reason:     { type: string, nullable: true }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    NetworkSecurityGroupRule:
      type: object
      required: [direction, protocol, port_range, cidr, action]
      properties:
        direction:  { type: string, enum: [ingress, egress] }
        protocol:   { type: string, enum: [tcp, udp, icmp, all] }
        port_range: { type: string, example: "443" }
        cidr:       { type: string, example: "0.0.0.0/0" }
        action:     { type: string, enum: [allow, deny] }
    NetworkSecurityGroup:
      type: object
      required: [id, tenant_id, name, rules, state, created_at, updated_at]
      properties:
        id:          { type: string }
        tenant_id:   { type: string }
        name:        { type: string }
        description: { type: string, nullable: true }
        rules:
          type: array
          items: { $ref: '#/components/schemas/NetworkSecurityGroupRule' }
        state:      { $ref: '#/components/schemas/NetworkResourceState' }
        reason:     { type: string, nullable: true }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    NetworkLoadBalancer:
      type: object
      required: [id, tenant_id, name, vpc_id, scheme, listeners, state, created_at, updated_at]
      properties:
        id:         { type: string }
        tenant_id:  { type: string }
        name:       { type: string }
        vpc_id:     { type: string }
        subnet_id:  { type: string, nullable: true }
        scheme:     { type: string, enum: [internal, public] }
        vip:        { type: string, nullable: true }
        listeners:
          type: array
          items: { $ref: '#/components/schemas/NetworkLoadBalancerListener' }
        state:      { $ref: '#/components/schemas/NetworkResourceState' }
        reason:     { type: string, nullable: true }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    NetworkVPCListResponse:
      type: object
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/NetworkVPC' } }
        total: { type: integer }
    NetworkSubnetListResponse:
      type: object
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/NetworkSubnet' } }
        total: { type: integer }
    NetworkSecurityGroupListResponse:
      type: object
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/NetworkSecurityGroup' } }
        total: { type: integer }
    NetworkLoadBalancerListResponse:
      type: object
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/NetworkLoadBalancer' } }
        total: { type: integer }
    CreateNetworkVPCRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 128 }
        cidr: { type: string, default: "10.0.0.0/16" }
    CreateNetworkSubnetRequest:
      type: object
      required: [vpc_id, name]
      properties:
        vpc_id:  { type: string }
        name:    { type: string, minLength: 1, maxLength: 128 }
        cidr:    { type: string, default: "10.0.1.0/24" }
        gateway: { type: string, nullable: true }
    CreateNetworkSecurityGroupRequest:
      type: object
      required: [name]
      properties:
        name:        { type: string, minLength: 1, maxLength: 128 }
        description: { type: string, nullable: true }
        rules:
          type: array
          items: { $ref: '#/components/schemas/NetworkSecurityGroupRule' }
    CreateNetworkLoadBalancerRequest:
      type: object
      required: [name, vpc_id]
      properties:
        name:      { type: string, minLength: 1, maxLength: 128 }
        vpc_id:    { type: string }
        subnet_id: { type: string, nullable: true }
        scheme:    { type: string, enum: [internal, public], default: internal }
        listeners:
          type: array
          items: { $ref: '#/components/schemas/NetworkLoadBalancerListener' }
...
  responses:
    Unauthorized:
      description: 未认证或 Token 无效（code=UNAUTHORIZED）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: 无权访问该资源（code=FORBIDDEN）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: 资源不存在（code=NOT_FOUND）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    BadRequest:
      description: 请求参数不合法（code=BAD_REQUEST）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# ── API 路径定义 ──────────────────────────────────────────────────────────────
# 说明：/demo/* 路径已迁移至 api/openapi/dev.yaml（不参与 SDK 生成）
paths:
...
  /networks/vpcs:
    get:
      operationId: listNetworkVPCs
      summary: 查询 VPC 列表
      tags: [Networks]
      x-ani-rbac-scope: "scope:networks:read"
      responses:
        "200":
          description: VPC 列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NetworkVPCListResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createNetworkVPC
      summary: 创建 VPC
      tags: [Networks]
      x-ani-rbac-scope: "scope:networks:create"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateNetworkVPCRequest' }
      responses:
        "201":
          description: VPC 已创建
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NetworkVPC' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /networks/vpcs/{vpc_id}:
    get:
      operationId: getNetworkVPC
      summary: 查询 VPC
      tags: [Networks]
      x-ani-rbac-scope: "scope:networks:read"
      parameters:
        - { name: vpc_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: VPC
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NetworkVPC' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteNetworkVPC
      summary: 删除 VPC
      tags: [Networks]
      x-ani-rbac-scope: "scope:networks:delete"
      parameters:
        - { name: vpc_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: VPC 已删除
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NetworkVPC' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
  /networks/subnets:
    get:
      operationId: listNetworkSubnets
      summary: 查询子网列表
      tags: [Networks]
      x-ani-rbac-scope: "scope:networks:read"
      responses:
        "200":
          description: 子网列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NetworkSubnetListResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createNetworkSubnet
      summary: 创建子网
      tags: [Networks]
      x-ani-rbac-scope: "scope:networks:create"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateNetworkSubnetRequest' }
      responses:
        "201":
          description: 子网已创建
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NetworkSubnet' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
  /networks/subnets/{subnet_id}:
    get:
      operationId: getNetworkSubnet
      summary: 查询子网
      tags: [Networks]
      x-ani-rbac-scope: "scope:networks:read"
      parameters:
        - { name: subnet_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: 子网
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NetworkSubnet' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteNetworkSubnet
      summary: 删除子网
      tags: [Networks]
      x-ani-rbac-scope: "scope:networks:delete"
      parameters:
        - { name: subnet_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: 子网已删除
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NetworkSubnet' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
  /networks/security-groups:
    get:
      operationId: listNetworkSecurityGroups
      summary: 查询安全组列表
      tags: [Networks]
      x-ani-rbac-scope: "scope:networks:read"
      responses:
        "200":
          description: 安全组列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NetworkSecurityGroupListResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createNetworkSecurityGroup
      summary: 创建安全组
      tags: [Networks]
      x-ani-rbac-scope: "scope:networks:create"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateNetworkSecurityGroupRequest' }
      responses:
        "201":
          description: 安全组已创建
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NetworkSecurityGroup' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /networks/security-groups/{security_group_id}:
    get:
      operationId: getNetworkSecurityGroup
      summary: 查询安全组
      tags: [Networks]
      x-ani-rbac-scope: "scope:networks:read"
      parameters:
        - { name: security_group_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: 安全组
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NetworkSecurityGroup' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteNetworkSecurityGroup
      summary: 删除安全组
      tags: [Networks]
      x-ani-rbac-scope: "scope:networks:delete"
      parameters:
        - { name: security_group_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: 安全组已删除
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NetworkSecurityGroup' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
  /networks/load-balancers:
    get:
      operationId: listNetworkLoadBalancers
      summary: 查询负载入口列表
      tags: [Networks]
      x-ani-rbac-scope: "scope:networks:read"
      responses:
        "200":
          description: 负载入口列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NetworkLoadBalancerListResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createNetworkLoadBalancer
      summary: 创建负载入口
      tags: [Networks]
      x-ani-rbac-scope: "scope:networks:create"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateNetworkLoadBalancerRequest' }
      responses:
        "201":
          description: 负载入口已创建
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NetworkLoadBalancer' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
  /networks/load-balancers/{load_balancer_id}:
    get:
      operationId: getNetworkLoadBalancer
      summary: 查询负载入口
      tags: [Networks]
      x-ani-rbac-scope: "scope:networks:read"
      parameters:
        - { name: load_balancer_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: 负载入口
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NetworkLoadBalancer' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteNetworkLoadBalancer
      summary: 删除负载入口
      tags: [Networks]
      x-ani-rbac-scope: "scope:networks:delete"
      parameters:
        - { name: load_balancer_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: 负载入口已删除
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NetworkLoadBalancer' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }

  # ── 存储资源（Core API Alpha）──────────────────────────────────────────────
...